requirements.txt
aiogram>=3.5
SQLAlchemy>=2.0
asyncpg>=0.29
aiosqlite>=0.20
python-dotenv>=1.0
